+++ test/roots/test-apidoc/lua/mod/inner.lua
@@ -1,0 +1,14 @@
+--- Description.
+local inner = {}
+
+--- Description.
+--- @type integer
+INNER_GLOBAL = 0
+
+--- Description.
+function inner.bar() end
+
+--- Description.
+--- @class mod.inner.Bar
+
+return inner