+++ test/roots/test-autodoc-emmylua/lua/incongruent_export.lua
@@ -1,7 +1,6 @@
 --- Description
 ---
 --- @class incongruent_export_2
---- @doctype module
 local incongruent_export = {}
 
 --- Description